Remove the watcher module.
